# api/swagger/swagger.yaml
basePath: /api
info:
  contact:
    email: abdoelrachmad@gmail.com
    name: Maintainer
  description: Simple blog server built with go.
  title: GoBlog Server
  version: "1.0"
paths:
...
  /v1/auth/writer/post:
    post:
      consumes:
      - application/json
      - application/msgpack
      description: Create a new post.
      parameters:
      - description: Create post form
        in: body
        name: form
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              categories:
                items:
                  type: string
                type: array
              content:
                type: string
              description:
                type: string
              featuringImagePath:
                type: string
              publishNow:
                type: boolean
              slug:
                type: string
              tags:
                items:
                  type: string
                type: array
              title:
                type: string
            type: object
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  allOf:
                  - type: object
                  - properties:
                      author:
                        allOf:
                        - type: object
                        - properties:
                            email:
                              type: string
                            firstName:
                              type: string
                            lastName:
                              type: string
                            uid:
                              type: string
                            username:
                              type: string
                          type: object
                      categories:
                        items:
                          allOf:
                          - type: object
                          - properties:
                              name:
                                type: string
                              slug:
                                type: string
                              uid:
                                type: string
                            type: object
                        type: array
                      commentCount:
                        type: integer
                      createdAt:
                        type: string
                      deletedAt:
                        type: string
                      description:
                        type: string
                      featuringImagePath:
                        type: string
                      publishedAt:
                        type: string
                      slug:
                        type: string
                      tags:
                        items:
                          type: string
                        type: array
                      title:
                        type: string
                      uid:
                        type: string
                      updatedAt:
                        type: string
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create Post
      tags:
      - Post (Writer)
...
  /v1/auth/writer/post/{uid}/depublish:
    patch:
      description: Remove publish status from my post if post already published.
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Unpublish My Post
      tags:
      - Post (Writer)
    put:
      description: Remove publish status from my post if post already published.
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Unpublish My Post
      tags:
      - Post (Writer)
  /v1/auth/writer/post/{uid}/detrash:
    patch:
      description: Restore my deleted post (soft-deleted).
      parameters:
      - description: Post's UID or slug
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Restore My Post (Soft)
      tags:
      - Post (Writer)
    put:
      description: Restore my deleted post (soft-deleted).
      parameters:
      - description: Post's UID or slug
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Restore My Post (Soft)
      tags:
      - Post (Writer)
  /v1/auth/writer/post/{uid}/permanent:
    delete:
      description: Delete my post (permanent).
      parameters:
      - description: Post's UID or slug
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete My Post (Permanent)
      tags:
      - Post (Writer)
  /v1/auth/writer/post/{uid}/publish:
    patch:
      description: Publish my post if not published yet.
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Publish My Post
      tags:
      - Post (Writer)
    put:
      description: Publish my post if not published yet.
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Publish My Post
      tags:
      - Post (Writer)
...
  /v1/auth/writer/posts/stats:
    get:
      description: Get my posts's stats.
      parameters:
      - description: Number of data to be shown.
        in: query
        name: show
        type: integer
      - description: Selected page of data.
        in: query
        name: page
        type: integer
      - description: Selected field to order data with.
        in: query
        name: order
        type: string
      - description: 'Ascending or descending, e.g.: ?asc=false.'
        in: query
        name: asc
        type: string
      - description: 'Filter data by type, e.g.: ?type=trash, ?type=published, ?type=draft.'
        in: query
        name: type
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  allOf:
                  - type: object
                  - properties:
                      currentPage:
                        type: integer
                      itemsPerPage:
                        type: integer
                      totalItems:
                        type: integer
                      totalPages:
                        type: integer
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get My Posts Stats
      tags:
      - Post (Writer)
...
securityDefinitions:
  BearerAuth:
    description: '`Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"